+++ functions.cpp
@@ -1,6 +1,6 @@
 ﻿#include "Header.h"
 
-void two_arrays(float num, int uah[], int pen[]) {
+void two_arrays(double num, int uah[], int pen[]) {
     long long int d = num * 100;
     int i = 9;
     int j = 9;
@@ -22,7 +22,7 @@ void two_arrays(float num, int uah[], int pen[]) {
     return;
 }
 
-void rename(int num, const int arr[]) {
+void rename(double num, const int arr[]) {
 	bool notNull = false;
 	if (num == 0) {
 		printf("нуль ");
